fix warnings
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@21931 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-common/src/main/java/com/tc/util/MaskingObjectIDSet.java
@@ -194,7 +194,7 @@ class MaskingObjectIDSet extends BitSetObjectIDSet {
 
     @Override
     public int size() {
-        return (int)((((int)(end-start)) - super.size()) + left.size());
+    return ((((int) (end - start)) - super.size()) + left.size());
     }
     
     private float factor() {
